+++ packages/eyes-cypress/CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Unreleased
 
+- fixed typescript definitions file for eyesCheckWindow string argument and open browser name parameter. 
 - updated VGC for using source instead of referer
 - updated to @applitools/eyes-sdk-core@10.1.2
-- updated to @applitools/visual-grid-client@14.0.0
+- updated to @applitools/visual-grid-client@14.0.1
 
 ## 3.10.4 - 2020/5/7