+++ packages/eyes-cypress/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+
+ ## 3.10.5 - 2020/5/18
+
 - fixed typescript definitions file for eyesCheckWindow string argument and open browser name parameter. 
 - updated to @applitools/eyes-sdk-core@10.1.2
 - updated to @applitools/visual-grid-client@14.0.2